Reimplimented default sort behaviour, cleaned up logs
