+++ autoscheduler/frontend/src/components/SchedulingPage/CourseSelectColumn/CourseSelectCard/ExpandedCourseCard/ExpandedCourseCard.tsx
@@ -31,8 +31,6 @@ const ExpandedCourseCard: React.FC<ExpandedCourseCardProps> = ({
   const dispatch = useDispatch();
   const { course, customizationLevel, loading } = courseCardOptions;
 
-  console.log('ExpandedCourseCard render()');
-
   const [options, setOptions] = React.useState([]);
   const [inputValue, setInputValue] = React.useState('');
